+++ cli/cli/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+### Added
+- New Code Analyzer to return compile time error for async void Callable methods.
+- New Code Fixer to fix async void Callable methods on IDE.
+- New Code Analyzer to validate Federations.
+- New Code Fixer to Implement possible fixes for Federations.
+- New Code Fixer to Solve Microservice classes missing Attribute or Partial keyword. 
+
 ## [4.2.0] - 2025-04-04
 ### Changed
 - `beam deploy` commands use solution level building instead of per-project [#3952](https://github.com/beamable/BeamableProduct/issues/3952)